pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trace-x-backend"
version = "0.1.0"
description = "Blockchain transaction analysis system for tracking fund flows and cross-chain transactions with integrated AML risk scoring"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "flasgger>=0.9.7",
    "etherscan-python>=2.1.0",
    "web3>=5.29,<6",
    "eth-abi>=2.2.0,<3.0.0",
    "eth-utils==1.10",
    "urllib3<2",
    "requests>=2.28.0",
    "numpy>=1.26.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "networkx>=3.3",
    "scikit-learn>=1.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-migrate>=4.0.0",
    "pymysql>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
src = ["configs/*.json"]
